move model templates to model folders, organize test chat sample across tests
