+++ tests/torchtune/data/test_chat_formats.py
@@ -6,34 +6,8 @@
 
 
 import pytest
-from tests.test_utils import assert_dialogue_equal
+from tests.test_utils import assert_dialogue_equal, MESSAGE_SAMPLE
 from torchtune.data import ChatMLFormat, Llama2ChatFormat, Message, MistralChatFormat
-
-# Taken from Open-Orca/SlimOrca-Dedup on HuggingFace:
-# https://huggingface.co/datasets/Open-Orca/SlimOrca-Dedup
-CHAT_SAMPLE = [
-    Message(
-        role="system",
-        content="You are an AI assistant. User will you give you a task. "
-        "Your goal is to complete the task as faithfully as you can. "
-        "While performing the task think step-by-step and justify your steps.",
-    ),
-    Message(
-        role="user",
-        content="Please briefly summarize this news article:\n\nAOL.com Video - "
-        "Father Lets 8-Year-Old Drive On Icy Road\n\nDescription:Would you let your "
-        "8-year-old drive your car? How about on an icy road? Well one father in "
-        "Russia did just that, and recorded the entire thing. To her credit, the "
-        "child seemed to be doing a great job. (0:44)\n\nTags: 8-year-old driver , "
-        "caught on camera , child driver , pix11\n\nSummary:",
-    ),
-    Message(
-        role="assistant",
-        content="A father in Russia allowed his 8-year-old child to drive his car "
-        "on an icy road and recorded the event. The child appeared to be handling the "
-        "situation well, showcasing their driving skills despite the challenging conditions.",
-    ),
-]
 
 
 class TestLlama2ChatFormat:
@@ -59,7 +33,7 @@ class TestLlama2ChatFormat:
     ]
 
     def test_format(self):
-        actual = Llama2ChatFormat.format(CHAT_SAMPLE)
+        actual = Llama2ChatFormat.format(MESSAGE_SAMPLE)
         assert_dialogue_equal(actual, self.expected_dialogue)
 
 
@@ -83,7 +57,7 @@ class TestMistralChatFormat:
     ]
 
     def test_format(self):
-        no_system_sample = CHAT_SAMPLE[1:]
+        no_system_sample = MESSAGE_SAMPLE[1:]
         actual = MistralChatFormat.format(no_system_sample)
         assert_dialogue_equal(actual, self.expected_dialogue)
 
@@ -91,7 +65,7 @@ class TestMistralChatFormat:
         with pytest.raises(
             ValueError, match="System prompts are not supported in MistralChatFormat"
         ):
-            _ = MistralChatFormat.format(CHAT_SAMPLE)
+            _ = MistralChatFormat.format(MESSAGE_SAMPLE)
 
 
 class TestChatMLFormat:
@@ -121,5 +95,5 @@ class TestChatMLFormat:
     ]
 
     def test_format(self):
-        actual = ChatMLFormat.format(CHAT_SAMPLE)
+        actual = ChatMLFormat.format(MESSAGE_SAMPLE)
         assert_dialogue_equal(actual, self.expected_dialogue)
