+++ tests/torchtune/data/test_messages.py
@@ -5,46 +5,18 @@
 # LICENSE file in the root directory of this source tree.
 
 import pytest
-from tests.test_utils import assert_dialogue_equal
+from tests.test_utils import (
+    assert_dialogue_equal,
+    CHAT_SAMPLE,
+    MESSAGE_SAMPLE,
+    MESSAGE_SAMPLE_TRAIN_ON_INPUT,
+)
 from torchtune.data._messages import (
     InputOutputToMessages,
     JSONToMessages,
     Message,
     ShareGPTToMessages,
 )
-
-
-# Taken from Open-Orca/SlimOrca-Dedup on Hugging Face:
-# https://huggingface.co/datasets/Open-Orca/SlimOrca-Dedup
-CHAT_SAMPLE = {
-    "system": "You are an AI assistant. User will you give you a task. Your goal is to complete the task as faithfully as you can. While performing the task think step-by-step and justify your steps.",  # noqa: B950
-    "user": "Please briefly summarize this news article:\n\nAOL.com Video - Father Lets 8-Year-Old Drive On Icy Road\n\nDescription:Would you let your 8-year-old drive your car? How about on an icy road? Well one father in Russia did just that, and recorded the entire thing. To her credit, the child seemed to be doing a great job. (0:44)\n\nTags: 8-year-old driver , caught on camera , child driver , pix11\n\nSummary:",  # noqa: B950
-    "assistant": "A father in Russia allowed his 8-year-old child to drive his car on an icy road and recorded the event. The child appeared to be handling the situation well, showcasing their driving skills despite the challenging conditions.",  # noqa: B950
-}
-
-EXPECTED_MESSAGE_TRAIN_ON_INPUT = [
-    Message(
-        role="system",
-        content=CHAT_SAMPLE["system"],
-    ),
-    Message(
-        role="user",
-        content=CHAT_SAMPLE["user"],
-    ),
-    Message(
-        role="assistant",
-        content=CHAT_SAMPLE["assistant"],
-    ),
-]
-
-EXPECTED_MESSAGE = [
-    Message(role="system", content=CHAT_SAMPLE["system"], masked=True),
-    Message(role="user", content=CHAT_SAMPLE["user"], masked=True),
-    Message(
-        role="assistant",
-        content=CHAT_SAMPLE["assistant"],
-    ),
-]
 
 
 class TestMessage:
@@ -154,13 +126,13 @@ class TestShareGPTToMessages:
     def test_call(self):
         transform = ShareGPTToMessages()
         converted_messages = transform(self.samples)
-        assert_dialogue_equal(converted_messages["messages"], EXPECTED_MESSAGE)
+        assert_dialogue_equal(converted_messages["messages"], MESSAGE_SAMPLE)
 
     def test_call_train_on_input(self):
         transform = ShareGPTToMessages(train_on_input=True)
         converted_messages = transform(self.samples)
         assert_dialogue_equal(
-            converted_messages["messages"], EXPECTED_MESSAGE_TRAIN_ON_INPUT
+            converted_messages["messages"], MESSAGE_SAMPLE_TRAIN_ON_INPUT
         )
 
 
@@ -185,11 +157,11 @@ class TestJSONToMessages:
     def test_call(self):
         transform = JSONToMessages()
         converted_messages = transform(self.samples)
-        assert_dialogue_equal(converted_messages["messages"], EXPECTED_MESSAGE)
+        assert_dialogue_equal(converted_messages["messages"], MESSAGE_SAMPLE)
 
     def test_call_train_on_input(self):
         transform = JSONToMessages(train_on_input=True)
         converted_messages = transform(self.samples)
         assert_dialogue_equal(
-            converted_messages["messages"], EXPECTED_MESSAGE_TRAIN_ON_INPUT
+            converted_messages["messages"], MESSAGE_SAMPLE_TRAIN_ON_INPUT
         )
