+++ tests/torchtune/models/llama2/test_llama2_prompt_template.py
@@ -1,0 +1,36 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+from tests.test_utils import assert_dialogue_equal, MESSAGE_SAMPLE
+from torchtune.data import Message
+from torchtune.models.llama2 import Llama2ChatTemplate
+
+
+class TestLlama2ChatTemplate:
+    expected_dialogue = [
+        Message(
+            role="user",
+            content="[INST] <<SYS>>\nYou are an AI assistant. User will you give you a task. "
+            "Your goal is to complete the task as faithfully as you can. While performing "
+            "the task think step-by-step and justify your steps.\n<</SYS>>\n\nPlease "
+            "briefly summarize this news article:\n\nAOL.com Video - Father Lets 8-Year-Old "
+            "Drive On Icy Road\n\nDescription:Would you let your 8-year-old drive your car? "
+            "How about on an icy road? Well one father in Russia did just that, and recorded "
+            "the entire thing. To her credit, the child seemed to be doing a great job. "
+            "(0:44)\n\nTags: 8-year-old driver , caught on camera , child driver , pix11\n\n"
+            "Summary: [/INST] ",
+        ),
+        Message(
+            role="assistant",
+            content="A father in Russia allowed his 8-year-old child to drive his car on an "
+            "icy road and recorded the event. The child appeared to be handling the situation well, "
+            "showcasing their driving skills despite the challenging conditions.",
+        ),
+    ]
+
+    def test_call(self):
+        actual = Llama2ChatTemplate()(MESSAGE_SAMPLE)
+        assert_dialogue_equal(actual, self.expected_dialogue)
